реструктурировал проект
lambda-frela/parser#10
